+++ app/src/main/java/com/emikhalets/medialib/data/entity/movies/MovieDetailsResponse.kt
@@ -6,7 +6,7 @@ import com.google.gson.annotations.SerializedName
 data class MovieDetailsResponse(
     @SerializedName("id") val id: Int? = null,
     @SerializedName("title") val title: String? = null,
-    @SerializedName("budget") val budget: Int? = null,
+    @SerializedName("budget") val budget: Long? = null,
     @SerializedName("backdrop_path") val backdrop: String? = null,
     @SerializedName("genres") val genres: List<GenreData>? = null,
     @SerializedName("imdb_id") val imdbId: String? = null,
@@ -14,7 +14,7 @@ data class MovieDetailsResponse(
     @SerializedName("overview") val overview: String? = null,
     @SerializedName("poster_path") val poster: String? = null,
     @SerializedName("release_date") val releaseDate: String? = null,
-    @SerializedName("revenue") val revenue: Int? = null,
+    @SerializedName("revenue") val revenue: Long? = null,
     @SerializedName("runtime") val runtime: Int? = null,
     @SerializedName("status") val status: String? = null,
     @SerializedName("tagline") val tagline: String? = null,
